chanstate: move revocation kv serialization
Move revocation log stream serialization and deserialization helpers
into the revocation KV store file. This keeps the pure revocation log
types separate from the on-disk encoding helpers while preserving the
existing function bodies and comments.
